--- src-ts/tools/work/work-list-title-renderer/index.ts
@@ -1,1 +1,0 @@
-export { default as WorkListTitleRenderer } from './WorkListTitleRenderer'
