+++ src-ts/tools/work/work-table-title-renderer/WorkTableTitleRenderer.module.scss
@@ -1,4 +1,4 @@
-.work-list-title-container {
+.work-table-title-container {
     display: flex;
     align-items: center;
 }
