+++ src-ts/tools/work/work-table-title-renderer/index.ts
@@ -1,0 +1,1 @@
+export { default as WorkTableTitleRenderer } from './WorkTableTitleRenderer'
