+++ src/routes/MyWork/components/WorkList/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { WorkList as WorkListTs } from '../../../../../src-ts/tools/work'
+import { WorkTable } from '../../../../../src-ts/tools/work'
 import { ProfileProvider, WorkProvider } from "../../../../../src-ts/lib"
 /**
  * Displays a list of work items for dashboard.
@@ -11,7 +11,7 @@ const WorkList = () => {
   return (
     <ProfileProvider>
       <WorkProvider>
-        <WorkListTs />
+        <WorkTable />
       </WorkProvider>
     </ProfileProvider>
   )
